backend/requirements.txt
fastapi
uvicorn
asyncio
python-dotenv
hume[stream]
manim
moviepy
openai
pydub